initial commandline rewrite extracted some commands to their own files
