+++ frontend/src/ts/elements/commandline-lists.ts
@@ -3,148 +3,14 @@ import * as Misc from "../utils/misc";
 import * as Notifications from "./notifications";
 import * as Sound from "../controllers/sound-controller";
 import * as ThemeController from "../controllers/theme-controller";
-import * as CustomTextPopup from "../popups/custom-text-popup";
-import * as ManualRestart from "../test/manual-restart-tracker";
 import Config, * as UpdateConfig from "../config";
 import * as PractiseWords from "../test/practise-words";
-import * as TestUI from "../test/test-ui";
 import * as TestLogic from "../test/test-logic";
 import * as Funbox from "../test/funbox";
-import * as TagController from "../controllers/tag-controller";
-import * as PresetController from "../controllers/preset-controller";
-import * as CustomText from "../test/custom-text";
-import * as Settings from "../pages/settings";
 import * as ChallengeController from "../controllers/challenge-controller";
-import * as PaceCaret from "../test/pace-caret";
-import * as TestInput from "../test/test-input";
-import * as ModesNotice from "../elements/modes-notice";
-import * as ConfigEvent from "../observables/config-event";
-import * as ShareTestSettingsPopup from "../popups/share-test-settings-popup";
+import * as ModesNotice from "./modes-notice";
 import { Auth } from "../firebase";
-import * as EditPresetPopup from "../popups/edit-preset-popup";
-import * as EditTagPopup from "../popups/edit-tags-popup";
 import { navigate } from "../controllers/route-controller";
-import * as TestState from "../test/test-state";
-import * as VideoAdPopup from "../popups/video-ad-popup";
-
-export let current: MonkeyTypes.CommandsGroup[] = [];
-
-function canBailOut(): boolean {
-  return (
-    (Config.mode === "custom" &&
-      CustomText.isWordRandom &&
-      (CustomText.word >= 5000 || CustomText.word == 0)) ||
-    (Config.mode === "custom" &&
-      !CustomText.isWordRandom &&
-      !CustomText.isTimeRandom &&
-      CustomText.text.length >= 5000) ||
-    (Config.mode === "custom" &&
-      CustomText.isTimeRandom &&
-      (CustomText.time >= 3600 || CustomText.time == 0)) ||
-    (Config.mode === "words" && Config.words >= 5000) ||
-    Config.words === 0 ||
-    (Config.mode === "time" && (Config.time >= 3600 || Config.time === 0)) ||
-    Config.mode == "zen"
-  );
-}
-
-const commandsLayouts: MonkeyTypes.CommandsGroup = {
-  title: "Layout emulator...",
-  configKey: "layout",
-  list: [
-    {
-      id: "couldnotload",
-      display: "Could not load the layouts list :(",
-    },
-  ],
-};
-
-Misc.getLayoutsList().then((layouts) => {
-  commandsLayouts.list = [];
-  commandsLayouts.list.push({
-    id: "changeLayoutDefault",
-    display: "off",
-    configValue: "default",
-    exec: (): void => {
-      UpdateConfig.setLayout("default");
-      TestLogic.restart();
-    },
-  });
-  Object.keys(layouts).forEach((layout) => {
-    commandsLayouts.list.push({
-      id: "changeLayout" + Misc.capitalizeFirstLetterOfEachWord(layout),
-      display: layout === "default" ? "off" : layout.replace(/_/g, " "),
-      configValue: layout,
-      exec: (): void => {
-        // UpdateConfig.setSavedLayout(layout);
-        UpdateConfig.setLayout(layout);
-        TestLogic.restart();
-      },
-    });
-  });
-});
-
-export const commandsKeymapLayouts: MonkeyTypes.CommandsGroup = {
-  title: "Change keymap layout...",
-  configKey: "keymapLayout",
-  list: [
-    {
-      id: "couldnotload",
-      display: "Could not load the layouts list :(",
-    },
-  ],
-};
-Misc.getLayoutsList().then((layouts) => {
-  commandsKeymapLayouts.list = [];
-  commandsKeymapLayouts.list.push({
-    id: "changeKeymapLayoutOverrideSync",
-    display: "emulator sync",
-    configValue: "overrideSync",
-    exec: (): void => {
-      UpdateConfig.setKeymapLayout("overrideSync");
-      TestLogic.restart();
-    },
-  });
-  Object.keys(layouts).forEach((layout) => {
-    if (layout.toString() != "default") {
-      commandsKeymapLayouts.list.push({
-        id: "changeKeymapLayout" + Misc.capitalizeFirstLetterOfEachWord(layout),
-        display: layout.replace(/_/g, " "),
-        configValue: layout,
-        exec: (): void => {
-          UpdateConfig.setKeymapLayout(layout);
-          TestLogic.restart();
-        },
-      });
-    }
-  });
-});
-
-const commandsLanguages: MonkeyTypes.CommandsGroup = {
-  title: "Language...",
-  configKey: "language",
-  list: [
-    {
-      id: "couldnotload",
-      display: "Could not load the languages list :(",
-    },
-  ],
-};
-
-commandsLanguages.list = [];
-Misc.getLanguageList().then((languages) => {
-  languages.forEach((language) => {
-    commandsLanguages.list.push({
-      id: "changeLanguage" + Misc.capitalizeFirstLetterOfEachWord(language),
-      display: language.replace(/_/g, " "),
-      configValue: language,
-      exec: (): void => {
-        UpdateConfig.setLanguage(language);
-        TestLogic.restart();
-      },
-    });
-  });
-});
 
 const commandsFunbox: MonkeyTypes.CommandsGroup = {
   title: "Funbox...",
@@ -215,135 +81,6 @@ Misc.getFontsList().then((fonts) => {
     },
   });
 });
-
-const commandsTags: MonkeyTypes.CommandsGroup = {
-  title: "Change tags...",
-  list: [],
-};
-
-export function updateTagCommands(): void {
-  const snapshot = DB.getSnapshot();
-  commandsTags.list = [];
-  if (!snapshot || !snapshot.tags || snapshot.tags.length === 0) {
-    commandsTags.list.push({
-      id: "createTag",
-      display: "Create tag",
-      icon: "fa-plus",
-      shouldFocusTestUI: false,
-      exec: (): void => {
-        EditTagPopup.show("add");
-      },
-    });
-    return;
-  }
-  commandsTags.list.push({
-    id: "clearTags",
-    display: `Clear tags`,
-    icon: "fa-times",
-    exec: (): void => {
-      const snapshot = DB.getSnapshot();
-
-      snapshot.tags = snapshot.tags?.map((tag) => {
-        tag.active = false;
-
-        return tag;
-      });
-
-      DB.setSnapshot(snapshot);
-      ModesNotice.update();
-      TagController.saveActiveToLocalStorage();
-    },
-  });
-
-  DB.getSnapshot().tags?.forEach((tag) => {
-    let dis = tag.display;
-
-    if (tag.active === true) {
-      dis = '<i class="fas fa-fw fa-check"></i>' + dis;
-    } else {
-      dis = '<i class="fas fa-fw"></i>' + dis;
-    }
-
-    commandsTags.list.push({
-      id: "toggleTag" + tag._id,
-      noIcon: true,
-      display: dis,
-      sticky: true,
-      exec: (): void => {
-        TagController.toggle(tag._id);
-        ModesNotice.update();
-
-        if (Config.paceCaret === "average") {
-          PaceCaret.init();
-          ModesNotice.update();
-        }
-
-        let txt = tag.display;
-
-        if (tag.active === true) {
-          txt = '<i class="fas fa-fw fa-check"></i>' + txt;
-        } else {
-          txt = '<i class="fas fa-fw"></i>' + txt;
-        }
-        if ($("#commandLine").hasClass("allCommands")) {
-          $(
-            `#commandLine .suggestions .entry[command='toggleTag${tag._id}']`
-          ).html(
-            `<div class="icon"><i class="fas fa-fw fa-tag"></i></div><div>Tags  > ` +
-              txt
-          );
-        } else {
-          $(
-            `#commandLine .suggestions .entry[command='toggleTag${tag._id}']`
-          ).html(txt);
-        }
-      },
-    });
-  });
-  commandsTags.list.push({
-    id: "createTag",
-    display: "Create tag",
-    icon: "fa-plus",
-    shouldFocusTestUI: false,
-    exec: (): void => {
-      EditTagPopup.show("add");
-    },
-  });
-}
-
-const commandsPresets: MonkeyTypes.CommandsGroup = {
-  title: "Presets...",
-  list: [],
-};
-
-export function updatePresetCommands(): void {
-  const snapshot = DB.getSnapshot();
-  commandsPresets.list = [];
-  if (!snapshot || !snapshot.presets || snapshot.presets.length === 0) return;
-  snapshot.presets.forEach((preset: MonkeyTypes.Preset) => {
-    const dis = preset.display;
-
-    commandsPresets.list.push({
-      id: "applyPreset" + preset._id,
-      display: dis,
-      exec: (): void => {
-        Settings.setEventDisabled(true);
-        PresetController.apply(preset._id);
-        Settings.setEventDisabled(false);
-        Settings.update();
-        ModesNotice.update();
-      },
-    });
-  });
-  commandsPresets.list.push({
-    id: "createPreset",
-    display: "Create preset",
-    icon: "fa-plus",
-    exec: (): void => {
-      EditPresetPopup.show("add");
-    },
-  });
-}
 
 const commandsRepeatQuotes: MonkeyTypes.CommandsGroup = {
   title: "Repeat quotes...",
@@ -1972,174 +1709,6 @@ const commandsTimerOpacity: MonkeyTypes.CommandsGroup = {
   ],
 };
 
-const commandsWordCount: MonkeyTypes.CommandsGroup = {
-  title: "Change word count...",
-  configKey: "words",
-  list: [
-    {
-      id: "changeWordCount10",
-      display: "10",
-      configValue: 10,
-      exec: (): void => {
-        UpdateConfig.setMode("words");
-        UpdateConfig.setWordCount(10);
-        TestLogic.restart();
-      },
-    },
-    {
-      id: "changeWordCount25",
-      display: "25",
-      configValue: 25,
-      exec: (): void => {
-        UpdateConfig.setMode("words");
-        UpdateConfig.setWordCount(25);
-        TestLogic.restart();
-      },
-    },
-    {
-      id: "changeWordCount50",
-      display: "50",
-      configValue: 50,
-      exec: (): void => {
-        UpdateConfig.setMode("words");
-        UpdateConfig.setWordCount(50);
-        TestLogic.restart();
-      },
-    },
-    {
-      id: "changeWordCount100",
-      display: "100",
-      configValue: 100,
-      exec: (): void => {
-        UpdateConfig.setMode("words");
-        UpdateConfig.setWordCount(100);
-        TestLogic.restart();
-      },
-    },
-    {
-      id: "changeWordCount200",
-      display: "200",
-      configValue: 200,
-      exec: (): void => {
-        UpdateConfig.setMode("words");
-        UpdateConfig.setWordCount(200);
-        TestLogic.restart();
-      },
-    },
-    {
-      id: "changeWordCountCustom",
-      display: "custom...",
-      input: true,
-      exec: (input): void => {
-        if (!input) return;
-        UpdateConfig.setMode("words");
-        UpdateConfig.setWordCount(parseInt(input));
-        TestLogic.restart();
-      },
-    },
-  ],
-};
-
-const commandsQuoteLengthConfig: MonkeyTypes.CommandsGroup = {
-  title: "Change quote length...",
-  configKey: "quoteLength",
-  list: [
-    {
-      id: "changeQuoteLengthAll",
-      display: "all",
-      configValue: [0, 1, 2, 3],
-      exec: (): void => {
-        UpdateConfig.setMode("quote");
-        UpdateConfig.setQuoteLength([0, 1, 2, 3]);
-        TestLogic.restart();
-      },
-    },
-    {
-      id: "changeQuoteLengthShort",
-      display: "short",
-      configValue: 0,
-      configValueMode: "include",
-      exec: (): void => {
-        UpdateConfig.setMode("quote");
-        UpdateConfig.setQuoteLength(0);
-        TestLogic.restart();
-      },
-    },
-    {
-      id: "changeQuoteLengthMedium",
-      display: "medium",
-      configValue: 1,
-      configValueMode: "include",
-      exec: (): void => {
-        UpdateConfig.setMode("quote");
-        UpdateConfig.setQuoteLength(1);
-        TestLogic.restart();
-      },
-    },
-    {
-      id: "changeQuoteLengthLong",
-      display: "long",
-      configValue: 2,
-      configValueMode: "include",
-      exec: (): void => {
-        UpdateConfig.setMode("quote");
-        UpdateConfig.setQuoteLength(2);
-        TestLogic.restart();
-      },
-    },
-    {
-      id: "changeQuoteLengthThicc",
-      display: "thicc",
-      configValue: 3,
-      configValueMode: "include",
-      exec: (): void => {
-        UpdateConfig.setMode("quote");
-        UpdateConfig.setQuoteLength(3);
-        TestLogic.restart();
-      },
-    },
-    {
-      id: "changeQuoteLengthFavorite",
-      display: "favorite",
-      configValue: -3,
-      configValueMode: "include",
-      available: (): boolean => {
-        return !!Auth.currentUser;
-      },
-      exec: (): void => {
-        UpdateConfig.setMode("quote");
-        UpdateConfig.setQuoteLength(-3);
-        TestLogic.restart();
-      },
-    },
-  ],
-};
-
-const commandsPunctuation: MonkeyTypes.CommandsGroup = {
-  title: "Change punctuation...",
-  configKey: "punctuation",
-  list: [
-    {
-      id: "changePunctuationOn",
-      display: "on",
-      configValue: true,
-      exec: (): void => {
-        UpdateConfig.setPunctuation(true);
-        TestLogic.restart();
-      },
-    },
-    {
-      id: "changePunctuationOff",
-      display: "off",
-      configValue: false,
-      exec: (): void => {
-        UpdateConfig.setPunctuation(false);
-        TestLogic.restart();
-      },
-    },
-  ],
-};
-
 const commandsNumbers: MonkeyTypes.CommandsGroup = {
   title: "Numbers...",
   configKey: "numbers",
@@ -2214,117 +1783,6 @@ const commandsQuickRestart: MonkeyTypes.CommandsGroup = {
       configValue: "off",
       exec: (): void => {
         UpdateConfig.setQuickRestartMode("off");
-      },
-    },
-  ],
-};
-
-const commandsMode: MonkeyTypes.CommandsGroup = {
-  title: "Change mode...",
-  configKey: "mode",
-  list: [
-    {
-      id: "changeModeTime",
-      display: "time",
-      configValue: "time",
-      exec: (): void => {
-        UpdateConfig.setMode("time");
-        TestLogic.restart();
-      },
-    },
-    {
-      id: "changeModeWords",
-      display: "words",
-      configValue: "words",
-      exec: (): void => {
-        UpdateConfig.setMode("words");
-        TestLogic.restart();
-      },
-    },
-    {
-      id: "changeModeQuote",
-      display: "quote",
-      configValue: "quote",
-      exec: (): void => {
-        UpdateConfig.setMode("quote");
-        TestLogic.restart();
-      },
-    },
-    {
-      id: "changeModeCustom",
-      display: "custom",
-      configValue: "custom",
-      exec: (): void => {
-        UpdateConfig.setMode("custom");
-        TestLogic.restart();
-      },
-    },
-    {
-      id: "changeModeZen",
-      display: "zen",
-      configValue: "zen",
-      exec: (): void => {
-        UpdateConfig.setMode("zen");
-        ManualRestart.set();
-        TestLogic.restart();
-      },
-    },
-  ],
-};
-
-const commandsTimeConfig: MonkeyTypes.CommandsGroup = {
-  title: "Change time config...",
-  configKey: "time",
-  list: [
-    {
-      id: "changeTimeConfig15",
-      display: "15",
-      configValue: 15,
-      exec: (): void => {
-        UpdateConfig.setMode("time");
-        UpdateConfig.setTimeConfig(15);
-        TestLogic.restart();
-      },
-    },
-    {
-      id: "changeTimeConfig30",
-      display: "30",
-      configValue: 30,
-      exec: (): void => {
-        UpdateConfig.setMode("time");
-        UpdateConfig.setTimeConfig(30);
-        TestLogic.restart();
-      },
-    },
-    {
-      id: "changeTimeConfig60",
-      display: "60",
-      configValue: 60,
-      exec: (): void => {
-        UpdateConfig.setMode("time");
-        UpdateConfig.setTimeConfig(60);
-        TestLogic.restart();
-      },
-    },
-    {
-      id: "changeTimeConfig120",
-      display: "120",
-      configValue: 120,
-      exec: (): void => {
-        UpdateConfig.setMode("time");
-        UpdateConfig.setTimeConfig(120);
-        TestLogic.restart();
-      },
-    },
-    {
-      id: "changeTimeConfigCustom",
-      display: "custom...",
-      input: true,
-      exec: (input): void => {
-        if (!input) return;
-        UpdateConfig.setMode("time");
-        UpdateConfig.setTimeConfig(parseInt(input));
-        TestLogic.restart();
       },
     },
   ],
@@ -2623,833 +2081,6 @@ const commandsMonkeyPowerLevel: MonkeyTypes.CommandsGroup = {
   ],
 };
 
-export const defaultCommands: MonkeyTypes.CommandsGroup = {
-  title: "",
-  list: [
-    {
-      id: "changePunctuation",
-      display: "Punctuation...",
-      icon: "!?",
-      subgroup: commandsPunctuation,
-    },
-    {
-      id: "changeMode",
-      display: "Mode...",
-      icon: "fa-bars",
-      subgroup: commandsMode,
-    },
-    {
-      id: "changeTimeConfig",
-      display: "Time...",
-      icon: "fa-clock",
-      subgroup: commandsTimeConfig,
-    },
-    {
-      id: "changeWordCount",
-      display: "Words...",
-      alias: "words",
-      icon: "fa-font",
-      subgroup: commandsWordCount,
-    },
-    {
-      id: "changeQuoteLength",
-      display: "Quote length...",
-      icon: "fa-quote-right",
-      alias: "quotes",
-      subgroup: commandsQuoteLengthConfig,
-    },
-    {
-      visible: false,
-      id: "changeTags",
-      display: "Tags...",
-      icon: "fa-tag",
-      subgroup: commandsTags,
-      beforeSubgroup: (): void => {
-        updateTagCommands();
-      },
-      available: (): boolean => {
-        return !!Auth.currentUser;
-      },
-      // exec: (): void => {
-      //   updateTagCommands();
-      //   current.push();
-      //   Commandline.show();
-      // },
-    },
-    {
-      visible: false,
-      id: "applyPreset",
-      display: "Presets...",
-      icon: "fa-sliders-h",
-      subgroup: commandsPresets,
-      beforeSubgroup: (): void => {
-        updatePresetCommands();
-      },
-      available: (): boolean => {
-        return !!Auth.currentUser;
-      },
-      // exec: (): void => {
-      //   updatePresetCommands();
-      //   current.push(commandsPresets);
-      //   Commandline.show();
-      // },
-    },
-    {
-      id: "changeConfidenceMode",
-      display: "Confidence mode...",
-      icon: "fa-backspace",
-      subgroup: commandsConfidenceMode,
-    },
-    {
-      id: "changeStopOnError",
-      display: "Stop on error...",
-      icon: "fa-hand-paper",
-      subgroup: commandsStopOnError,
-    },
-    {
-      id: "changeNumbers",
-      display: "Numbers...",
-      icon: "15",
-      subgroup: commandsNumbers,
-    },
-    {
-      id: "changeSmoothCaret",
-      display: "Smooth caret...",
-      icon: "fa-i-cursor",
-      subgroup: commandsSmoothCaret,
-    },
-    {
-      id: "changeQuickRestart",
-      display: "Quick restart...",
-      icon: "fa-redo-alt",
-      subgroup: commandsQuickRestart,
-    },
-    {
-      id: "changeRepeatQuotes",
-      display: "Repeat quotes...",
-      icon: "fa-sync-alt",
-      subgroup: commandsRepeatQuotes,
-    },
-    {
-      id: "changeLiveWpm",
-      display: "Live WPM...",
-      icon: "fa-tachometer-alt",
-      subgroup: commandsLiveWpm,
-    },
-    {
-      id: "changeLiveAcc",
-      display: "Live accuracy...",
-      icon: "fa-percentage",
-      subgroup: commandsLiveAcc,
-    },
-    {
-      id: "changeLiveBurst",
-      display: "Live burst...",
-      icon: "fa-fire-alt",
-      subgroup: commandsLiveBurst,
-    },
-    {
-      id: "changeShowTimer",
-      display: "Timer/progress...",
-      icon: "fa-clock",
-      subgroup: commandsShowTimer,
-    },
-    {
-      id: "changeKeyTips",
-      display: "Key tips...",
-      icon: "fa-question",
-      subgroup: commandsKeyTips,
-    },
-    {
-      id: "changeFreedomMode",
-      display: "Freedom mode...",
-      subgroup: commandsFreedomMode,
-    },
-    {
-      id: "changeStrictSpace",
-      display: "Strict space...",
-      icon: "fa-minus",
-      subgroup: commandsStrictSpace,
-    },
-    {
-      id: "changeBlindMode",
-      display: "Blind mode...",
-      icon: "fa-eye-slash",
-      subgroup: commandsBlindMode,
-    },
-    {
-      id: "changeShowWordsHistory",
-      display: "Always show words history...",
-      icon: "fa-align-left",
-      subgroup: commandsShowWordsHistory,
-    },
-    {
-      id: "changeIndicateTypos",
-      display: "Indicate typos...",
-      icon: "fa-exclamation",
-      subgroup: commandsIndicateTypos,
-    },
-    {
-      id: "changeHideExtraLetters",
-      display: "Hide extra letters...",
-      icon: "fa-eye-slash",
-      subgroup: commandsHideExtraLetters,
-    },
-    {
-      id: "changeQuickEnd",
-      display: "Quick end...",
-      icon: "fa-step-forward",
-      subgroup: commandsQuickEnd,
-    },
-    {
-      id: "singleListCommandLine",
-      display: "Single list command line...",
-      icon: "fa-list",
-      subgroup: commandsSingleListCommandLine,
-    },
-    {
-      id: "capsLockWarning",
-      display: "Caps lock warning...",
-      icon: "fa-exclamation-triangle",
-      subgroup: commandsCapsLockWarning,
-    },
-    {
-      id: "changeMinWpm",
-      display: "Minimum wpm...",
-      alias: "minimum",
-      icon: "fa-bomb",
-      subgroup: commandsMinWpm,
-    },
-    {
-      id: "changeMinAcc",
-      display: "Minimum accuracy...",
-      alias: "minimum",
-      icon: "fa-bomb",
-      subgroup: commandsMinAcc,
-    },
-    {
-      id: "changeMinBurst",
-      display: "Minimum burst...",
-      alias: "minimum",
-      icon: "fa-bomb",
-      subgroup: commandsMinBurst,
-    },
-    {
-      id: "changeOppositeShiftMode",
-      display: "Change opposite shift mode...",
-      icon: "fa-exchange-alt",
-      subgroup: commandsOppositeShiftMode,
-    },
-    {
-      id: "changeSoundOnClick",
-      display: "Sound on click...",
-      icon: "fa-volume-up",
-      subgroup: commandsSoundOnClick,
-    },
-    {
-      id: "changeSoundOnError",
-      display: "Sound on error...",
-      icon: "fa-volume-mute",
-      subgroup: commandsSoundOnError,
-    },
-    {
-      id: "changeSoundVolume",
-      display: "Sound volume...",
-      icon: "fa-volume-down",
-      subgroup: commandsSoundVolume,
-    },
-    {
-      id: "changeFlipTestColors",
-      display: "Flip test colors...",
-      icon: "fa-adjust",
-      subgroup: commandsFlipTestColors,
-    },
-    {
-      id: "changeSmoothLineScroll",
-      display: "Smooth line scroll...",
-      icon: "fa-align-left",
-      subgroup: commandsSmoothLineScroll,
-    },
-    {
-      id: "changeAlwaysShowDecimal",
-      display: "Always show decimal places...",
-      icon: "00",
-      subgroup: commandsAlwaysShowDecimal,
-    },
-    {
-      id: "changeAlwaysShowCPM",
-      display: "Always show CPM...",
-      icon: "fa-tachometer-alt",
-      subgroup: commandsAlwaysShowCPM,
-    },
-    {
-      id: "changeStartGraphsAtZero",
-      display: "Start graphs at zero...",
-      icon: "fa-chart-line",
-      subgroup: commandsStartGraphsAtZero,
-    },
-    {
-      id: "changeLazyMode",
-      display: "Lazy mode...",
-      icon: "fa-couch",
-      subgroup: commandsLazyMode,
-    },
-    {
-      id: "changeShowAllLines",
-      display: "Show all lines...",
-      icon: "fa-align-left",
-      subgroup: commandsShowAllLines,
-    },
-    {
-      id: "changeColorfulMode",
-      display: "Colorful mode...",
-      icon: "fa-fill-drip",
-      subgroup: commandsColorfulMode,
-    },
-    {
-      id: "changeOutOfFocusWarning",
-      display: "Out of focus warning...",
-      icon: "fa-exclamation",
-      subgroup: commandsOutOfFocusWarning,
-    },
-    {
-      id: "setEnableAds",
-      display: "Enable ads...",
-      icon: "fa-ad",
-      subgroup: commandsEnableAds,
-    },
-    {
-      id: "changeTheme",
-      display: "Theme...",
-      icon: "fa-palette",
-      subgroup: themeCommands,
-    },
-    {
-      id: "setCustomTheme",
-      display: "Custom theme...",
-      icon: "fa-palette",
-      subgroup: customThemeCommands,
-    },
-    {
-      id: "setCustomThemeId",
-      display: "Custom themes...",
-      icon: "fa-palette",
-      subgroup: customThemeListCommands,
-      beforeSubgroup: (): void => updateCustomThemeListCommands(),
-      available: (): boolean => {
-        return Auth.currentUser !== null;
-      },
-    },
-    {
-      id: "changeRandomTheme",
-      display: "Random theme...",
-      icon: "fa-random",
-      subgroup: commandsRandomTheme,
-    },
-    {
-      id: "randomizeTheme",
-      display: "Next random theme",
-      icon: "fa-random",
-      exec: (): Promise<void> => ThemeController.randomizeTheme(),
-      available: (): boolean => {
-        return Config.randomTheme !== "off";
-      },
-    },
-    {
-      id: "changeDifficulty",
-      display: "Difficulty...",
-      icon: "fa-star",
-      subgroup: commandsDifficulty,
-    },
-    {
-      id: "changeCaretStyle",
-      display: "Caret style...",
-      icon: "fa-i-cursor",
-      subgroup: commandsCaretStyle,
-    },
-    {
-      id: "changePaceCaret",
-      display: "Pace caret mode...",
-      icon: "fa-i-cursor",
-      subgroup: commandsPaceCaret,
-    },
-    {
-      id: "changePaceCaretStyle",
-      display: "Pace caret style...",
-      icon: "fa-i-cursor",
-      subgroup: commandsPaceCaretStyle,
-    },
-    {
-      id: "changeRepeatedPace",
-      display: "Repeated pace...",
-      icon: "fa-i-cursor",
-      subgroup: commandsRepeatedPace,
-    },
-    {
-      id: "changeTimerStyle",
-      display: "Timer/progress style...",
-      icon: "fa-clock",
-      subgroup: commandsTimerStyle,
-    },
-    {
-      id: "changeTimerColor",
-      display: "Timer/progress color...",
-      icon: "fa-clock",
-      subgroup: commandsTimerColor,
-    },
-    {
-      id: "changeTimerOpacity",
-      display: "Timer/progress opacity...",
-      icon: "fa-clock",
-      subgroup: commandsTimerOpacity,
-    },
-    {
-      id: "changeHighlightMode",
-      display: "Highlight mode...",
-      icon: "fa-highlighter",
-      subgroup: commandsHighlightMode,
-    },
-    {
-      id: "changeTapeMode",
-      display: "Tape mode...",
-      icon: "fa-tape",
-      subgroup: commandsTapeMode,
-    },
-    {
-      id: "changeShowAverage",
-      display: "Show average...",
-      icon: "fa-chart-bar",
-      subgroup: commandsShowAverage,
-    },
-    {
-      id: "changeCustomBackground",
-      display: "Custom background...",
-      icon: "fa-image",
-      defaultValue: "",
-      input: true,
-      exec: (input): void => {
-        if (!input) input = "";
-        UpdateConfig.setCustomBackground(input);
-      },
-    },
-    {
-      id: "changeLanguage",
-      display: "Language...",
-      icon: "fa-language",
-      subgroup: commandsLanguages,
-    },
-    {
-      id: "changeBritishEnglish",
-      display: "British english...",
-      icon: "fa-language",
-      subgroup: commandsBritishEnglish,
-    },
-    {
-      id: "changeFunbox",
-      display: "Funbox...",
-      alias: "fun box",
-      icon: "fa-gamepad",
-      subgroup: commandsFunbox,
-    },
-    {
-      id: "changeLayout",
-      display: "Layout emulator...",
-      icon: "fa-keyboard",
-      subgroup: commandsLayouts,
-    },
-    {
-      id: "toggleKeymap",
-      display: "Keymap mode...",
-      icon: "fa-keyboard",
-      alias: "keyboard",
-      subgroup: commandsKeymapMode,
-    },
-    {
-      id: "changeKeymapStyle",
-      display: "Keymap style...",
-      alias: "keyboard",
-      icon: "fa-keyboard",
-      subgroup: commandsKeymapStyle,
-    },
-    {
-      id: "changeKeymapLegendStyle",
-      display: "Keymap legend style...",
-      alias: "keyboard",
-      icon: "fa-keyboard",
-      subgroup: commandsKeymapLegendStyle,
-    },
-    {
-      id: "changeKeymapLayout",
-      display: "Keymap layout...",
-      alias: "keyboard",
-      icon: "fa-keyboard",
-      subgroup: commandsKeymapLayouts,
-    },
-    {
-      id: "changeKeymapShowTopRow",
-      display: "Keymap show top row...",
-      alias: "keyboard",
-      icon: "fa-keyboard",
-      subgroup: commandsKeymapShowTopRow,
-    },
-    {
-      id: "changeCustomLayoutfluid",
-      display: "Custom layoutfluid...",
-      defaultValue: "qwerty dvorak colemak",
-      input: true,
-      icon: "fa-tint",
-      exec: (input): void => {
-        if (input === undefined) return;
-        UpdateConfig.setCustomLayoutfluid(
-          input as MonkeyTypes.CustomLayoutFluidSpaces
-        );
-        if (Config.funbox === "layoutfluid") TestLogic.restart();
-        // UpdateConfig.setLayout(
-        //   Config.customLayoutfluid
-        //     ? Config.customLayoutfluid.split("_")[0]
-        //     : "qwerty"
-        // );
-        // UpdateConfig.setKeymapLayout(
-        //   Config.customLayoutfluid
-        //     ? Config.customLayoutfluid.split("_")[0]
-        //     : "qwerty"
-        // );
-      },
-    },
-    {
-      id: "changeFontSize",
-      display: "Font size...",
-      icon: "fa-font",
-      input: true,
-      exec: (input): void => {
-        if (!input) return;
-        UpdateConfig.setFontSize(parseFloat(input));
-        setTimeout(() => {
-          TestUI.updateWordsHeight();
-        }, 0); //honestly no clue why it i need to wait for the next event loop to do this
-      },
-    },
-    {
-      id: "changeFontFamily",
-      display: "Font family...",
-      icon: "fa-font",
-      subgroup: commandsFonts,
-    },
-    {
-      id: "changePageWidth",
-      display: "Page width...",
-      icon: "fa-arrows-alt-h",
-      subgroup: commandsPageWidth,
-    },
-    {
-      id: "nextTest",
-      display: "Next test",
-      alias: "restart start begin type test typing",
-      icon: "fa-chevron-right",
-      available: (): boolean => {
-        return TestUI.resultVisible;
-      },
-      exec: (): void => {
-        TestLogic.restart();
-      },
-    },
-    {
-      id: "viewTypingPage",
-      display: "View Typing Page",
-      alias: "start begin type test",
-      icon: "fa-keyboard",
-      exec: (): void => {
-        navigate("/");
-      },
-    },
-    {
-      id: "viewLeaderboards",
-      display: "View Leaderboards",
-      icon: "fa-crown",
-      exec: (): void => {
-        $("#top #menu .textButton.view-leaderboards").trigger("click");
-      },
-    },
-    {
-      id: "viewAbout",
-      display: "View About Page",
-      icon: "fa-info",
-      exec: (): void => {
-        navigate("/about");
-      },
-    },
-    {
-      id: "viewSettings",
-      display: "View Settings Page",
-      icon: "fa-cog",
-      exec: (): void => {
-        navigate("/settings");
-      },
-    },
-    {
-      id: "viewQuoteSearchPopup",
-      display: "Search for quotes",
-      icon: "fa-search",
-      exec: (): void => {
-        UpdateConfig.setMode("quote");
-        $("#quote-search-button").trigger("click");
-      },
-      shouldFocusTestUI: false,
-    },
-    {
-      id: "viewAccount",
-      display: "View Account Page",
-      icon: "fa-user",
-      alias: "stats",
-      exec: (): void => {
-        $("#top #menu .textButton.view-account").hasClass("hidden")
-          ? navigate("/login")
-          : navigate("/account");
-      },
-    },
-    {
-      id: "toggleFullscreen",
-      display: "Toggle Fullscreen",
-      icon: "fa-expand",
-      exec: (): void => {
-        Misc.toggleFullscreen();
-      },
-    },
-    {
-      id: "bailOut",
-      display: "Bail out...",
-      icon: "fa-running",
-      subgroup: {
-        title: "Are you sure...",
-        list: [
-          {
-            id: "bailOutNo",
-            display: "Nevermind",
-            available: (): boolean => {
-              return canBailOut();
-            },
-          },
-          {
-            id: "bailOutForSure",
-            display: "Yes, I am sure",
-            exec: (): void => {
-              TestInput.setBailout(true);
-              TestLogic.finish();
-            },
-            available: (): boolean => {
-              return canBailOut();
-            },
-          },
-        ],
-      },
-      visible: false,
-      available: (): boolean => {
-        return canBailOut();
-      },
-    },
-    {
-      id: "loadChallenge",
-      display: "Load challenge...",
-      icon: "fa-award",
-      subgroup: commandsChallenges,
-    },
-    {
-      id: "joinDiscord",
-      display: "Join the Discord server",
-      icon: "fa-users",
-      exec: (): void => {
-        window.open("https://discord.gg/monkeytype");
-      },
-    },
-    {
-      id: "repeatTest",
-      display: "Repeat test",
-      icon: "fa-sync-alt",
-      exec: (): void => {
-        TestLogic.restart({
-          withSameWordset: true,
-        });
-      },
-      available: (): boolean => {
-        return TestUI.resultVisible;
-      },
-    },
-    {
-      id: "practiseWords",
-      display: "Practice words...",
-      icon: "fa-exclamation-triangle",
-      subgroup: commandsPractiseWords,
-      available: (): boolean => {
-        return TestUI.resultVisible;
-      },
-    },
-    {
-      id: "toggleWordHistory",
-      display: "Toggle word history",
-      icon: "fa-align-left",
-      exec: (): void => {
-        TestUI.toggleResultWords();
-      },
-      available: (): boolean => {
-        return TestUI.resultVisible;
-      },
-    },
-    {
-      id: "saveScreenshot",
-      display: "Save screenshot",
-      icon: "fa-image",
-      alias: "ss picture",
-      exec: (): void => {
-        setTimeout(() => {
-          TestUI.screenshot();
-        }, 500);
-      },
-      available: (): boolean => {
-        return TestUI.resultVisible;
-      },
-    },
-    {
-      id: "changeCustomModeText",
-      display: "Change custom text",
-      icon: "fa-align-left",
-      exec: (): void => {
-        CustomTextPopup.show();
-      },
-    },
-    {
-      id: "toggleMonkey",
-      display: "Toggle Monkey",
-      icon: "fa-egg",
-      visible: false,
-      exec: (): void => {
-        UpdateConfig.setMonkey(!Config.monkey);
-      },
-    },
-    {
-      id: "copyWordsToClipboard",
-      display: "Copy words to clipboard",
-      icon: "fa-copy",
-      subgroup: true,
-      exec: (): void => {
-        current.push(commandsCopyWordsToClipboard);
-      },
-    },
-    {
-      id: "importSettingsJSON",
-      display: "Import settings JSON",
-      icon: "fa-cog",
-      alias: "import config",
-      input: true,
-      exec: (input): void => {
-        if (!input) return;
-        try {
-          UpdateConfig.apply(JSON.parse(input));
-          UpdateConfig.saveFullConfigToLocalStorage();
-          Settings.update();
-          Notifications.add("Done", 1);
-        } catch (e) {
-          Notifications.add(
-            "An error occured while importing settings: " + e,
-            -1
-          );
-        }
-      },
-    },
-    {
-      id: "exportSettingsJSON",
-      display: "Export settings JSON",
-      icon: "fa-cog",
-      alias: "export config",
-      input: true,
-      defaultValue: "",
-    },
-    {
-      id: "monkeyPower",
-      display: "Power mode...",
-      alias: "powermode",
-      icon: "fa-egg",
-      visible: false,
-      subgroup: commandsMonkeyPowerLevel,
-    },
-    {
-      id: "watchVideoAd",
-      display: "Watch video ad",
-      alias: "support donate",
-      icon: "fa-ad",
-      exec: (): void => {
-        VideoAdPopup.show();
-      },
-    },
-    {
-      id: "shareTestSettings",
-      display: "Share test settings",
-      icon: "fa-share",
-      exec: async (): Promise<void> => {
-        ShareTestSettingsPopup.show();
-      },
-    },
-    {
-      id: "setResultSaving",
-      display: "Result saving...",
-      icon: "fa-save",
-      alias: "results",
-      subgroup: {
-        title: "Result saving...",
-        list: [
-          {
-            id: "setResultSavingOff",
-            display: "off",
-            alias: "disabled",
-            exec: (): void => {
-              TestState.setSaving(false);
-              ModesNotice.update();
-            },
-          },
-          {
-            id: "setResultSavingOn",
-            display: "on",
-            alias: "enabled",
-            exec: (): void => {
-              TestState.setSaving(true);
-              ModesNotice.update();
-            },
-          },
-        ],
-      },
-    },
-    {
-      id: "clearSwCache",
-      display: "Clear SW cache",
-      icon: "fa-cog",
-      exec: async (): Promise<void> => {
-        const clist = await caches.keys();
-        for (const name of clist) {
-          await caches.delete(name);
-        }
-        window.location.reload();
-      },
-    },
-    {
-      id: "getSwCache",
-      display: "Get SW cache",
-      icon: "fa-cog",
-      exec: async (): Promise<void> => {
-        alert(await caches.keys());
-      },
-    },
-  ],
-};
-
-current = [defaultCommands];
-
-export function setCurrent(val: MonkeyTypes.CommandsGroup[]): void {
-  current = val;
-}
-
-export function pushCurrent(val: MonkeyTypes.CommandsGroup): void {
-  current.push(val);
-}
-
 const listsObject = {
   commandsChallenges,
   commandsLanguages,
@@ -3473,26 +2104,3 @@ export type ListsObjectKeys = keyof typeof listsObject;
 export function getList(list: ListsObjectKeys): MonkeyTypes.CommandsGroup {
   return listsObject[list];
 }
-
-ConfigEvent.subscribe((eventKey, eventValue) => {
-  if (eventKey === "saveToLocalStorage") {
-    defaultCommands.list.filter(
-      (command) => command.id == "exportSettingsJSON"
-    )[0].defaultValue = eventValue as string;
-  }
-  if (eventKey === "customBackground") {
-    defaultCommands.list.filter(
-      (command) => command.id == "changeCustomBackground"
-    )[0].defaultValue = eventValue as string;
-  }
-  if (eventKey === "fontSize") {
-    defaultCommands.list.filter(
-      (command) => command.id == "changeFontSize"
-    )[0].defaultValue = eventValue as string;
-  }
-  if (eventKey === "customLayoutFluid") {
-    defaultCommands.list.filter(
-      (command) => command.id == "changeCustomLayoutfluid"
-    )[0].defaultValue = (eventValue as string)?.replace(/#/g, " ");
-  }
-});
